refactor: ExceptionDto를 BasicResponse로 변경하여 재사용성 확보
