+++ springboot-jwt/src/main/java/com/example/springbootjwt/global/common/dto/BasicResponse.java
@@ -1,4 +1,4 @@
-package com.example.springbootjwt.global.common.exception;
+package com.example.springbootjwt.global.common.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @AllArgsConstructor
-public class ExceptionDto {
+public class BasicResponse {
     private String message;
     private HttpStatus status;
 }
